+++ mllopenai/__init__.py
@@ -1,1 +1,3 @@
-from .main import OpenAIManager
+from .main import ChatManager
+from .main import ImageManager
+from .main import SpeechManager